Handlers for onLoad and onError can have return value
The handlers given for onLoad and onError now can return a value. In that case, that value is passed on to the next listener. If null is returned, the original value is passed to the next listener.

Along with this, added a constructor that only takes a next listener. This is useful to make a listener null-proof.