Feature/random jira issues (#496)
* APPDEV-1044, APPDEV-1075, APPDEV-1074, APPDEV-1076, APPDEV-1069, APPDEV-1079

* APPDEV-1044 — Limited data fetched at the app start up. App will now fetch only data for days and weeks as directed by server url’s.
APPDEV-1075 — App is now listening to reboot and starts the background service to listen and post the app activities.
APPDEV-1069 — Environment switch options re-enables and made dynamic to take the user entered url.
APPDEV-1079 — App now uses to get the next and current data as per urls in the Links from response.

* Use lambda for listeners

Refactored ActivityCategoryManagerImpl to use the new DataLoadListenerImpl, which supports lambda expressions for listeners

* Handlers for onLoad and onError can have return value

The handlers given for onLoad and onError now can return a value. In that case, that value is passed on to the next listener. If null is returned, the original value is passed to the next listener.

Along with this, added a constructor that only takes a next listener. This is useful to make a listener null-proof.

* Move away from Jack

To do that, upgraded to Gradle 4.4 and Android plug-in 3.1.0

* Upgrade to Android plugin 3.1.3

Advised by Android Studio

* Code Refactored

-- Using Lambdas and handlers for errors.

* Code Refactoring

APPDEV-1069 -- Server-Environment-switch
	-- Updated localized string values.
	-- Xml file renamed .

Along with that broken long function in NotificationManagerImpl.
Removed empty lines.

* Code Refactoring.

-- Breaking down of big functions.

* Issue: Buddy Timeline is showing only one goal per day per each user post my commits.
Reason : There was a hidden break statement present in the code which is skipping the loop. (Added while exploring the functionality).
Along with this broke down the Long method (for buddy timeline response processing)  into sub methods.

* PR #496 Review changes.
 	-- Refactoring the long functions in Activities and Notifications implementations.
	-- Removed the duplicate functions which are adding page and size and modifying the url .
	-- Removed the duplicate entires in Notifications list.
	-- Added Yona message text to the title of the notifications in fragment.

* Addressed my own review comments

* Deduplicated some code
* Removed some unused methods
* Removed some commented-out code
* Corrected some spelling mistakes
* Removed some superfluous null checks