Feature/appdev 1119 build up on jenkins (#498)
* APPDEV-1044, APPDEV-1075, APPDEV-1074, APPDEV-1076, APPDEV-1069, APPDEV-1079

* APPDEV-1044 — Limited data fetched at the app start up. App will now fetch only data for days and weeks as directed by server url’s.
APPDEV-1075 — App is now listening to reboot and starts the background service to listen and post the app activities.
APPDEV-1069 — Environment switch options re-enables and made dynamic to take the user entered url.
APPDEV-1079 — App now uses to get the next and current data as per urls in the Links from response.

* Use lambda for listeners

Refactored ActivityCategoryManagerImpl to use the new DataLoadListenerImpl, which supports lambda expressions for listeners

* Handlers for onLoad and onError can have return value

The handlers given for onLoad and onError now can return a value. In that case, that value is passed on to the next listener. If null is returned, the original value is passed to the next listener.

Along with this, added a constructor that only takes a next listener. This is useful to make a listener null-proof.

* Move away from Jack

To do that, upgraded to Gradle 4.4 and Android plug-in 3.1.0

* Upgrade to Android plugin 3.1.3

Advised by Android Studio

* Code Refactored

-- Using Lambdas and handlers for errors.

* Code Refactoring

APPDEV-1069 -- Server-Environment-switch
	-- Updated localized string values.
	-- Xml file renamed .

Along with that broken long function in NotificationManagerImpl.
Removed empty lines.

* Code Refactoring.

-- Breaking down of big functions.

* Issue: Buddy Timeline is showing only one goal per day per each user post my commits.
Reason : There was a hidden break statement present in the code which is skipping the loop. (Added while exploring the functionality).
Along with this broke down the Long method (for buddy timeline response processing)  into sub methods.

* PR #496 Review changes.
 	-- Refactoring the long functions in Activities and Notifications implementations.
	-- Removed the duplicate functions which are adding page and size and modifying the url .
	-- Removed the duplicate entires in Notifications list.
	-- Added Yona message text to the title of the notifications in fragment.

* Addressed my own review comments

* Deduplicated some code
* Removed some unused methods
* Removed some commented-out code
* Corrected some spelling mistakes
* Removed some superfluous null checks

* created Jenkinstest

* corrected checkout url

* check for sdk manager

* check for sdk manager

* sdk manager check

* Android sdk update

* Android update to sdk 27

* Android skd upadate

* Android_home updated

* adding gradle build step

* print tools dir

* accept androdi licenses

* check for android tools

* ls android home

* android sdk ls

* printing the android platfroms

* added gradle step after updating android sdk from regrular build job

* / missing in gradle command

* increase ram for gradle

* --scan added gradle step

* Trying sdkmanager --licenses option

* Corrected quote

* Exploring directories

* Use android tool rather than sdkmanager

* Update SDK

* List bin64 folder

* Add y response for android update

* Add --no-daemon

To try and overcome a daemon crash

* Android-27 instead of 25

* Added steps back

* List tools folder again

* Add verbose option

* Moved verbose option to the right place

* Update Jenkinsfile

* Added build-tools-27.0.3

* Removed  -Dorg.gradle.jvmargs=-Xmx1536m

* Removed platform-tools,android-27,extra-android-m2repository

* Removed scan option for Gradle

* Removed --no-daemon again

Along with this chose the assemble target rather than assembleDebug

* Going back to assembleDebug

Because of this error:
Keystore file '/scratch/jenkins/workspace/ona-app-android_Jenkinstest-KMB7MK3VI3M7IKUEKTDG72QR57B3HKMBFTOY6NFN2OKA6LZVB3JA/app/null/yona.jks' not found for signing config 'release'.

* Corrected path for archiveArtifacts

Along with that switched allowEmptyArchive to false, as it's a failure to have an empty archive.

* Set build number from environment

For Jenkins builds, it will come from Jenkins. Local builds need to set the environment variable

* Different way to access environment

* Use BUILD_NUMBER as integer

* Single stage with tagging

* Corrected repo

* Added GIT credentials

* Put the branch name in the tag

* Removed People section

* Improved the definition statement

* Updated prerequisites

* Minor update

* printing env variables in parallel stage

* reverted the paralled stage

* reverting the changes made by pipeline editor.

* Added android-27 again

* Checking the git status

* APPDEV-1119 Explicit versionCode management

The version code is now stored in a branch-specific file, allowing to manage it specifically per branch.