APPDEV-1189 Send log messages to Crashlytics
Crashlytics supports adding log messages to the crash report, which could be of great help when analyzing issues. If Fabric is not initialized, the log messages are issued the classic way. This can occur in early stages of the app and in unit tests.

Log messages reported through reportException are now not logged anymore when they are being passed on to a listener. That's fair, as the listener is responsible for doing the handling of that error. The listener might log it or handle it otherwise.